Added hud, added player model, started work on shop

+++ IsoArcher/GoldManager/GlobalGoldManager/GlobalGoldManager.cs
@@ -4,5 +4,5 @@ using System;
 public static class GlobalGoldManager
 {
     // Players current gold
-    public static int globalGold = 0;
+    public static float globalGold = 0;
 }
